Removed various un-used files and made some style and also code-fixes

+++ src/posts/list-post/list-post.component.ts
@@ -13,6 +13,7 @@ import {User} from '../../users/models/user.model';
 export class ListPostComponent implements OnInit {
 
   posts: Post[];
+  confirmationMessage = 'Are you sure you want to delete this Post?';
 
   constructor(private route: ActivatedRoute, private postService: PostService, private userService: UserService, private router: Router) {
     this.posts = [];
@@ -39,19 +40,21 @@ export class ListPostComponent implements OnInit {
    * @param post - Post Object
    */
   deletePost(post: Post): void {
-    this.postService.deletePost(post).subscribe(() => {
-      // Replace and Update user-object
-      let elementIndex = 0;
-      this.posts.forEach((element, index) => {
-        if (element.id === post.id) {
-          elementIndex = index;
-        }
-      });
-      // Remove Element from users Array
-      this.posts.splice(elementIndex, 1);
+    if (confirm(this.confirmationMessage)) {
+      this.postService.deletePost(post).subscribe(() => {
+        // Replace and Update user-object
+        let elementIndex = 0;
+        this.posts.forEach((element, index) => {
+          if (element.id === post.id) {
+            elementIndex = index;
+          }
+        });
+        // Remove Element from users Array
+        this.posts.splice(elementIndex, 1);
 
-      alert('Successfully Removed the Post');
-    });
+        alert('Successfully Removed the Post');
+      });
+    }
   }
 
   ngOnInit(): void {
